refactor: decompose ConversationRuntime
